# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v1.3.1
servers:
  - url: http://localhost:8080
    description: local
  - url: /histudy
    description: dev
security:
  - bearerAuth: [ ]
paths:
...
  /api/users:
    get:
      tags:
        - 일반 사용자 API
      summary: 유저 검색
      operationId: searchUser
      parameters:
        - name: search
          in: query
          required: false
          allowEmptyValue: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
      deprecated: true
    post:
      tags:
        - 일반 사용자 API
      summary: 회원가입
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserForm'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserLogin'
  /api/v2/users:
    get:
      tags:
        - 일반 사용자 API
      summary: 유저 검색
      operationId: searchUserWithMasking
      parameters:
        - name: search
          in: query
          required: false
          allowEmptyValue: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
...
  /api/courses:
    get:
      tags:
        - 강의 관리 API
      summary: 강의 목록 조회
      operationId: getCourses
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
    post:
      tags:
        - 강의 관리 API
      summary: 강의 목록 업로드
      operationId: importCourses
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
...
  /api/team/reports:
    get:
      tags:
        - 스터디 그룹 API
      summary: 그룹 보고서 목록 조회
      operationId: getMyGroupReports
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReportDto'
    post:
      tags:
        - 스터디 그룹 API
      summary: 그룹 스터디 보고서 생성
      operationId: createReport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportForm'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReportInfo'
  /api/team/reports/{reportId}:
    get:
      tags:
        - 스터디 그룹 API
      summary: 그룹 특정 보고서 조회
      operationId: getReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReportInfo'
    delete:
      tags:
        - 스터디 그룹 API
      summary: 그룹 특정 보고서 삭제
      operationId: deleteReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
    patch:
      tags:
        - 스터디 그룹 API
      summary: 그룹 특정 보고서 수정
      operationId: updateReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportForm'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/team/reports/image:
    post:
      tags:
        - 스터디 그룹 API
      summary: 스터디 보고서에 들어갈 인증 이미지 업로드
      operationId: uploadImage
      parameters: [ ]
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                imagePath: /path/to/image.png
  /api/team/reports/{reportIdOr}/image:
    post:
      tags:
        - 스터디 그룹 API
      summary: 스터디 보고서에 들어갈 인증 이미지 업로드
      operationId: uploadImage_1
      parameters:
        - name: reportIdOr
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                imagePath: /path/to/image.png

  # ===================
  # Admin APIs
  # ===================
...
components:
  schemas:
...
    CourseInfo:
      type: object
      properties:
        id:
          type: number
          description: Course ID
          example: 1
        name:
          type: string
          description: Course Name
          example: Software Engineering
        prof:
          type: string
          description: Course Professor
          example: Prof. John Doe
        code:
          type: string
          description: Course Code
          example: CSEE 4111
        year:
          type: number
          description: Course Year
          example: 2021
        semester:
          type: number
          description: Course Semester
          example: 1
      description: List of course added to apply form
    UserBasic:
      type: object
      properties:
        id:
          type: number
          description: User ID
          example: 1
        name:
          type: string
          description: User name
          example: John Doe
        sid:
          type: string
          description: User student ID
          example: '211234567'
      description: List of friend added to apply form
    UserForm:
      type: object
      properties:
        sub:
          type: string
          description: User ID
          example: '1234567890'
        name:
          type: string
          description: User Name
          example: John Doe
        email:
          type: string
          description: User Email
          example: jd@example.com
        sid:
          type: string
          description: User Student ID
          example: '21800111'
    JwtPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      description: Token pairs
    UserLogin:
      type: object
      properties:
        isRegistered:
          type: boolean
          description: Registration status
          example: true
        tokenType:
          type: string
          description: Token type
          example: 'Bearer '
        tokens:
          $ref: '#/components/schemas/JwtPair'
        role:
          type: string
          description: Role
          example: USER
    ReportForm:
      type: object
      properties:
        title:
          type: string
          description: Report Title
          example: Week 15 Report
        content:
          type: string
          description: Report Content
          example: This is a report for week 15.
        totalMinutes:
          type: number
          description: Total minutes of the report
          example: 60
        participants:
          type: array
          description: Participant ID(PK) of the report
          example: [ 1, 2 ]
          items:
            type: integer
            description: Participant ID(PK) of the report
            format: int64
        images:
          type: array
          description: Image URLs of the report
          example: [ "/path/to/image1.png", "/path/to/image2.png" ]
          items:
            type: string
            description: Image URLs of the report
            example: '"/path/to/image1.png","/path/to/image2.png"'
        courses:
          type: array
          description: Course IDs of the report
          example: [ 1, 2 ]
          items:
            type: integer
            description: Course IDs of the report
            format: int64
    BasicCourseInfo:
      type: object
      properties:
        id:
          type: number
          description: Course ID
          example: 1
        name:
          type: string
          description: Course Name
          example: Software Engineering
        prof:
          type: string
          description: Course Professor
          example: Prof. John Doe
      description: Course names of the report
      example: [ OOP, OS ]
    ImageDto:
      type: object
      properties:
        id:
          type: number
          description: Image ID
          example: 1
        url:
          type: string
          description: Image URL
          example: /path/to/image.png
      description: Images of the report
    ReportInfo:
      type: object
      properties:
        id:
          type: number
          description: Report ID
          example: 1
        title:
          type: string
          description: Report Title
          example: Week 15 Report
        content:
          type: string
          description: Report Content
          example: This is a report for week 15
        totalMinutes:
          type: number
          description: Total minutes of the report
          example: 60
        participants:
          type: array
          description: Participant SIDs of the report
          items:
            $ref: '#/components/schemas/UserBasic'
        courses:
          type: array
          description: Course names of the report
          example: [ OOP, OS ]
          items:
            $ref: '#/components/schemas/BasicCourseInfo'
        images:
          type: array
          description: Images of the report
          items:
            $ref: '#/components/schemas/ImageDto'
        regDate:
          type: string
          description: Report Last Modified Date
          example: '2021-06-01 00:00:00'
...
    UserDto:
      type: object
      properties:
        users:
          type: array
          description: List of user
          items:
            $ref: '#/components/schemas/UserMatching'
    UserMatching:
      type: object
      properties:
        name:
          type: string
          description: User name
          example: John Doe
        sid:
          type: string
          description: User student ID
          example: '211234567'
        email:
          type: string
          description: User email
          example: jd@example.com
      description: List of user
...
    ReportDto:
      type: object
      properties:
        reports:
          type: array
          description: List of reports
          items:
            $ref: '#/components/schemas/ReportInfo'
    CourseDto:
      type: object
      properties:
        courses:
          type: array
          description: List of courses
          items:
            $ref: '#/components/schemas/CourseInfo'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT